Move API client into a submodule
